+++ Annuaire/src/main/java/fr/yannm/annuaire/model/person/CreatePerson.java
@@ -1,4 +1,4 @@
-package fr.yannm.annuaire.model;
+package fr.yannm.annuaire.model.person;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,8 +8,8 @@ import lombok.Setter;
 /**
  * @author Yann
  * @version 1.0
- * @name : UpdatePerson
- * @created 08/12/2021 - 11:15
+ * @name : CreatePerson
+ * @created 07/12/2021 - 17:01
  * @project Annuaire
  * @copyright Yann
  **/
@@ -17,7 +17,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdatePerson {
+public class CreatePerson {
 
     private String name;
     private String surname;
